pyproject.toml
[project]
name = "langchain-documentation-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4",
    "black",
    "tiktoken",
    "openai",
    "unstructured",
    "nltk",
    "fastapi",
    "jinja2",
    "uvicorn",
    "streamlit",
    "streamlit-chat",
    "tqdm",
    "isort",
    "langchainhub",
    "firecrawl-py",
    "langchain-community",
    "langsmith",
    "pillow",
    "certifi",
    "langchain-chroma",
    "python-dotenv",
    "langchain-tavily",
    "langchain-pinecone",
    "rich>=14.3.3",
]

[dependency-groups]
dev = [
    "ipykernel",
]
